+++ src/entities/comment/index.ts
@@ -1,0 +1,1 @@
+//TODO структура см. post
