+++ src/pages/auth_page/ui/auth_container.tsx
