+++ src/pages/auth_page/ui/sign_up_form.tsx
