+++ src/pages/post_view_page/api/get_post_by_id.ts
