+++ src/pages/post_view_page/ui/post_details.css
