+++ src/pages/post_view_page/ui/post_details.tsx
@@ -1,6 +1,6 @@
 import {useEffect, useState} from "preact/hooks";
-import type {Post} from "@/shared/models/post.ts";
-import {getPostById} from "@/pages/post_details/api/get_post_by_id.ts";
+import type {Post} from "@/entities/post/model/post.types";
+import {getPostById} from "../api/get_post_by_id";
 import {useRoute} from "preact-iso";
 import type {UUIDv4} from "@/shared/lib/uuid";
 import './post_details.css'
