+++ src/shared/assets/index.ts
@@ -1,0 +1,1 @@
+//TODO здесь будет статика
