+++ src/shared/config/index.ts
@@ -1,0 +1,1 @@
+//TODO здесь будут константы
