+++ src/shared/hooks/index.ts
@@ -1,0 +1,1 @@
+//TODO здесь будут переиспользуемые хуки
