+++ src/shared/ui/index.ts
@@ -1,0 +1,1 @@
+// TODO: здесь будут общие для всего приложения компоненты (кнопки, поля и т.п.)
